on_trial_end callback
