+++ js/event-listeners.js
@@ -1,12 +1,15 @@
-import { get_is_playing, set_is_playing, n_back_level} from "./state-manager.js";
-import {start_trial, stop_trial} from "./audio.js";
+import { get_is_playing, n_back_level} from "./state-manager.js";
+import {start_trial, stop_trial} from "./cues.js";
 import { check_audio_hit, check_visual_hit } from "./n-back.js";
 
 const start_stop_button = document.getElementById("start-stop");
+const audio_button = document.getElementById("audio-match");
+const visual_button = document.getElementById("visual-match");
+
+
 
 function add_start_stop_listener(){
     document.addEventListener("DOMContentLoaded", function() {
-        const start_stop_button = document.getElementById("start-stop");
         start_stop_button.addEventListener("click", () => {
             console.log("start/stop button clicked");
     
@@ -25,8 +28,7 @@ function add_start_stop_listener(){
 
 function add_audio_cue_listener(){
     document.addEventListener("DOMContentLoaded", function() {
-        const start_stop_button = document.getElementById("audio-match");
-        start_stop_button.addEventListener("click", () => {
+        audio_button.addEventListener("click", () => {
     
             check_audio_hit(n_back_level);
     
@@ -37,8 +39,7 @@ function add_audio_cue_listener(){
 
 function add_visual_cue_listener(){
     document.addEventListener("DOMContentLoaded", function() {
-        const start_stop_button = document.getElementById("visual-match");
-        start_stop_button.addEventListener("click", () => {
+        visual_button.addEventListener("click", () => {
     
             check_visual_hit(n_back_level);
