--- python-code/.vscode/launch.json
@@ -1,28 +1,0 @@
-{
-    // Use IntelliSense to learn about possible attributes.
-    // Hover to view descriptions of existing attributes.
-    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-    "version": "0.2.0",
-    "configurations": [
-        
-        {
-            "name": "Python: Retrieve-Audio",
-            "type": "python",
-            "request": "launch",
-            "module": "flask",
-            "python": "${workspaceFolder}/../.venv/bin/python",
-            "env": {
-                "FLASK_APP": "retrieve-audio-ms.py",
-                "FLASK_DEBUG": "0",
-                "FLASK_RUN_PORT": "8000"
-            },
-            "args": [
-                "run",
-                "--no-debugger",
-                "--no-reload"
-            ],
-            "jinja": true,
-            "justMyCode": true
-        }
-    ]
-}
