--- symlinks/apache-config.conf
@@ -1,1 +1,0 @@
-/etc/apache2/sites-enabled/000-default.conf
